# Process input files in parallel