Process input source files in parallel to reduce the total AST generation time. The time taken to generate an AST is cut down by more than half with parallel processing.

## About the changes
A [PLINQ query](https://learn.microsoft.com/en-us/dotnet/standard/parallel-programming/introduction-to-plinq) is used to parallelize the `foreach` loops that iterate over files in the input directory one by one.

Since the `AsParallel` method is called on an `IEnumerable<FileInfo>`, PLINQ should opt for chunk partitioning (dynamic batch sizes) by default, as opposed to a range partitioning (fixed batch sizes) approach as suggested in #11. 

Refs:
- https://learn.microsoft.com/en-us/dotnet/standard/parallel-programming/custom-partitioners-for-plinq-and-tpl#kinds-of-partitioning
- https://devblogs.microsoft.com/pfxteam/chunk-partitioning-vs-range-partitioning-in-plinq/

## Benchmark

A fresh clone of the [bitwarden/server](https://github.com/bitwarden/server) repo was created locally (at commit [`4adcecb`](https://github.com/bitwarden/server/commit/4adcecb80a980f9f9690c5f19e22072e61f5038d)).

Then, a scan was run against the cloned repo twice, once from the `main` branch of this repo and once with the code changes in this PR.

```shell
time dotnet run --project DotNetAstGen/ -i /tmp/bitwarden-server -o /tmp/bitwarden-server-ast/
```

The results of the scans are added below:

**Without the changes from this PR (sequential processing)**:
```
________________________________________________________
Executed in   88.05 secs    fish           external
   usr time   87.45 secs   49.00 micros   87.45 secs
   sys time    4.36 secs  662.00 micros    4.36 secs
```

**With the changes from this PR (parallel processing)**
```
________________________________________________________
Executed in   39.36 secs    fish           external
   usr time  158.99 secs    0.08 millis  158.99 secs
   sys time   10.21 secs    1.48 millis   10.21 secs
```

Thus, AST generation with parallel processing was 48.69 seconds or about 55% faster.

I did multiple runs and the result was similar each time.

For reference these runs were done on a machine running macOS 15.0.1 with Apple M3 Pro and 36GiB RAM.